refactor: rename protobuf.fixtures.js to applyOperation.fixtures.js and update imports
