+++ tests/unit/network/v1/V1BroadcastTransactionResponse.test.js
@@ -15,7 +15,7 @@ import {publicKeyToAddress} from '../../../../src/utils/helpers.js';
 import {OperationType, ResultCode} from '../../../../src/utils/constants.js';
 import { config } from '../../../helpers/config.js';
 import { createState } from '../utils/createState.js';
-import protobufFixtures from '../../../fixtures/protobuf.fixtures.js';
+import protobufFixtures from '../../../fixtures/applyOperation.fixtures.js';
 import { testKeyPair1 } from '../../../fixtures/apply.fixtures.js';
 
 const remotePublicKey = b4a.from(testKeyPair1.publicKey, 'hex');
